Split middleware in separate files, injected the hook as param
